Add full logic for displaying test result on the backend and frontend sides, add displaying results from url page and fix bug in top bar

+++ frontend/src/app/core/converters/time-converter.ts
@@ -3,6 +3,6 @@ export class TimeConverter {
         // time has format 'hh:mm:ss'
         const colons = time.split(':');
 
-        return parseInt(colons[0]) * 60 * 60 + parseInt(colons[1]) * 60 + parseInt(colons[2]);
+        return parseInt(colons[0]) * 60 * 60 + parseInt(colons[1]) * 60 + parseFloat(colons[2]);
     }
 }
